Move CreateTraceLocationApiV1.kt and TraceLocationServer.kt into new sub-package
