+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/eventregistration/events/server/createtracelocation/TraceLocationServer.kt
@@ -1,4 +1,4 @@
-package de.rki.coronawarnapp.eventregistration.events.server
+package de.rki.coronawarnapp.eventregistration.events.server.createtracelocation
 
 import androidx.annotation.VisibleForTesting
 import dagger.Lazy
